int64_t

+++ circle/src/lib/types/vector.c
@@ -3,37 +3,37 @@
 
 #include <stdlib.h>
 
-int *NewVector(int capacity)
+int64_t *NewVector(int64_t capacity)
 {
-    int *vec = mem_malloc((capacity + 2) * sizeof(int));
+    int64_t *vec = mem_malloc((capacity + 2) * sizeof(int64_t));
     vec[1] = capacity;
 
     return vec + 2;
 }
 
-void DestroyVector(int *v)
+void DestroyVector(int64_t *v)
 {
-    int *vec = v - 2;
+    int64_t *vec = v - 2;
     free(vec);
 }
 
-int Vector_GetLen(int *v)
+int Vector_GetLen(int64_t *v)
 {
-    int *vec = v - 2;
+    int64_t *vec = v - 2;
 
     return *vec;
 }
 
-void Vector_Push(int **v, const int a)
+void Vector_Push(int64_t **v, const int64_t a)
 {
-    int *vec = *v - 2;
+    int64_t *vec = *v - 2;
     const int len = *vec;
-    int capacity = *(vec + 1);
+    int64_t capacity = *(vec + 1);
 
     if (len == capacity)
     {
         capacity *= 2;
-        vec = mem_realloc(vec, (capacity + 2) * sizeof(int));
+        vec = mem_realloc(vec, (capacity + 2) * sizeof(int64_t));
         *(vec + 1) = capacity;
         *v = vec + 2;
     }
@@ -45,18 +45,18 @@ void Vector_Push(int **v, const int a)
 #include "test.h"
 void TEST_Vector()
 {
-    int *v = NewVector(2);
+    int64_t *v = NewVector(2);
     Vector_Push(&v, 2);
     if (*v != 2)
     {
-        FAIL("Can't push 2 to vector got %d", *v);
+        FAIL("Can't push 2 to vector got %lld", *v);
     }
     Vector_Push(&v, 3);
     Vector_Push(&v, 4);
 
     if (v[2] != 4)
     {
-        FAIL("Can't grow vector got %d", v[2]);
+        FAIL("Can't grow vector got %lld", v[2]);
     }
 
     DestroyVector(v);
